phpstan and refactor for level 0
