+++ .gitignore
@@ -2,4 +2,6 @@
 /tests/cache/*
 .git
 .phpunit.result.cache
-.phpunit.result.cache
+
+/build/
+.php-cs-fixer.cache
