+++ composer.json
@@ -12,7 +12,8 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.4.0",
-        "pestphp/pest": "1.21.2"
+        "pestphp/pest": "1.21.2",
+        "phpstan/phpstan": "1.6.x-dev"
     },
     "authors": [
         {
